feat: move navigation controls to title bar
